Move import-light module to the blocks folder
