+++ client/blocks/import-light/README.md
@@ -1,0 +1,8 @@
+# IMPORT LIGHT
+
+`import-light` is a block module that provides components for building import light flow (ex. onboarding flow):
+
+- `capture` - capturing site for import from
+- `scanning` - scanning screen
+- `colors` - list of captured colors
+- `summery` - summary screen showing all grabbed details from the source site
