+++ client/blocks/import-light/capture/style.scss
