+++ client/blocks/import-light/scanning/index.tsx
